feat: Renamed snapp references in Sudoku to zk
Renames all references of 'snapps' to either reference 'zk' or 'zkapps'
where appropriate. These changes have been tested by building and
running the tests. Left a TODO comment regarding the return value of
Mina.getAccount named as snapp. This should be addressed later.
